otel: move panic-hook test back to an integration test
Per review, crash telemetry is better covered as an integration test
than a module-local unit test, so move it back to
tests/otel_panic_hook.rs (reverting the earlier relocation into
src/otel.rs). This intentionally takes the documented exception to
tests/AGENTS.md's "real TestingServer, no mocks" rule — the hook fires
on process crash, which a TestingServer flow can't drive, and an
in-memory exporter is the closest faithful stand-in for the node
collector.

Adds the /// doc comment AGENTS.md requires (contract, actor, ASCII
flow), and keeps the self-contained hardening from the unit-test version
(local Arc sentinel, save/restore of the prior global panic hook).
