+++ crates/jazz-tools/src/otel.rs
@@ -260,69 +260,4 @@ mod tests {
             .expect("active_websockets gauge datapoint present");
         assert_eq!(observed, 1);
     }
-
-    // A panic must become a flushed, structured OTLP log record (carrying the
-    // message + source location), and the previously installed hook must still
-    // run so the stderr backtrace / filelog safety net survives. Uses an
-    // in-memory exporter so it needs no collector, and restores the prior
-    // global panic hook so it doesn't leak into the rest of the test binary.
-    #[test]
-    fn panic_hook_emits_flushed_otlp_log_and_preserves_previous_hook() {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicBool, Ordering};
-
-        use opentelemetry_sdk::logs::InMemoryLogExporter;
-        use tracing_subscriber::prelude::*;
-
-        let exporter = InMemoryLogExporter::default();
-        let provider = SdkLoggerProvider::builder()
-            .with_batch_exporter(exporter.clone())
-            .build();
-
-        // Route `tracing` events to the OTLP bridge for this thread. The panic
-        // hook runs on the panicking thread, so this thread-local default is
-        // active when the hook emits its `error!` event.
-        let subscriber = tracing_subscriber::registry()
-            .with(log_bridge::<tracing_subscriber::Registry>(&provider));
-        let _subscriber_guard = tracing::subscriber::set_default(subscriber);
-
-        // Sentinel previous hook (in a local Arc) — install_panic_hook must
-        // capture and re-invoke it. Save the prior global hook and restore it
-        // after the panic so this test doesn't disturb the rest of the binary.
-        let saved_hook = std::panic::take_hook();
-        let previous_ran = Arc::new(AtomicBool::new(false));
-        let previous_ran_in_hook = previous_ran.clone();
-        std::panic::set_hook(Box::new(move |_| {
-            previous_ran_in_hook.store(true, Ordering::SeqCst);
-        }));
-        install_panic_hook(provider.clone());
-
-        let result = std::panic::catch_unwind(|| panic!("boom-from-test"));
-        std::panic::set_hook(saved_hook);
-
-        assert!(result.is_err(), "the panic should have been caught");
-        assert!(
-            previous_ran.load(Ordering::SeqCst),
-            "install_panic_hook must re-invoke the previously installed hook",
-        );
-
-        let logs = exporter.get_emitted_logs().expect("read emitted logs");
-        let panic_record = logs
-            .iter()
-            .find(|log| format!("{:?}", log.record.body()).contains("process panicked"))
-            .expect("panic hook should emit a 'process panicked' log record");
-
-        let attrs = format!(
-            "{:?}",
-            panic_record.record.attributes_iter().collect::<Vec<_>>()
-        );
-        assert!(
-            attrs.contains("panic.message") && attrs.contains("boom-from-test"),
-            "panic log must carry the panic message as an attribute; got: {attrs}",
-        );
-        assert!(
-            attrs.contains("panic.location"),
-            "panic log must carry the source location; got: {attrs}",
-        );
-    }
 }
